fix: resolve import cycle
