+++ packages/runtime/src/helpers/config.ts
@@ -1,6 +1,6 @@
 import type { NetlifyConfig } from '@netlify/build'
 import destr from 'destr'
-import { readJSON, writeJSON } from 'fs-extra'
+import { readJSON } from 'fs-extra'
 import type { Header } from 'next/dist/lib/load-custom-routes'
 import type { NextConfigComplete } from 'next/dist/server/config-shared'
 import { join, dirname, relative } from 'pathe'
@@ -10,7 +10,8 @@ import glob from 'tiny-glob'
 import { HANDLER_FUNCTION_NAME, IMAGE_FUNCTION_NAME, ODB_FUNCTION_NAME } from '../constants'
 
 import { splitApiRoutes, useNoneBundler } from './flags'
-import { APILambda, getCommonDependencies } from './functions'
+import type { APILambda } from './functions'
+import { getCommonDependencies, getRequiredServerFiles } from './tracing'
 import type { RoutesManifest } from './types'
 import { escapeStringRegexp } from './utils'
 
@@ -55,24 +56,6 @@ export const getNextConfig = async function getNextConfig({
   } catch (error: unknown) {
     return failBuild('Error loading your Next config', { error })
   }
-}
-
-/**
- * Returns all of the NextJS configuration stored within 'required-server-files.json'
- * To update the configuration within this file, use the 'updateRequiredServerFiles' method.
- */
-export const getRequiredServerFiles = async (publish: string): Promise<RequiredServerFiles> => {
-  const configFile = join(publish, 'required-server-files.json')
-  return await readJSON(configFile)
-}
-
-/**
- * Writes a modified configuration object to 'required-server-files.json'.
- * To get the full configuration, use the 'getRequiredServerFiles' method.
- */
-export const updateRequiredServerFiles = async (publish: string, modifiedConfig: RequiredServerFiles) => {
-  const configFile = join(publish, 'required-server-files.json')
-  await writeJSON(configFile, modifiedConfig)
 }
 
 // hack to make files like `[id].js` work.
