+++ packages/runtime/src/helpers/functions.ts
@@ -5,7 +5,7 @@ import destr from 'destr'
 import { copyFile, ensureDir, existsSync, readJSON, writeFile, writeJSON, stat } from 'fs-extra'
 import type { ImageConfigComplete, RemotePattern } from 'next/dist/shared/lib/image-config'
 import { outdent } from 'outdent'
-import { join, relative, resolve, dirname } from 'pathe'
+import { join, relative, resolve } from 'pathe'
 import glob from 'tiny-glob'
 
 import {
@@ -22,10 +22,10 @@ import { getHandler } from '../templates/getHandler'
 import { getResolverForPages, getResolverForSourceFiles } from '../templates/getPageResolver'
 
 import { ApiConfig, extractConfigFromFile, isEdgeConfig } from './analysis'
-import { getRequiredServerFiles } from './config'
-import { getDependenciesOfFile, getServerFile, getSourceFileForPage } from './files'
+import { getServerFile, getSourceFileForPage } from './files'
 import { writeFunctionConfiguration } from './functionsMetaData'
 import { pack } from './pack'
+import { getCommonDependencies, getDependenciesOfFile } from './tracing'
 import { ApiRouteType } from './types'
 import { getFunctionNameForPage } from './utils'
 
@@ -226,74 +226,6 @@ export const setupImageFunction = async ({
       status: 200,
     })
   }
-}
-
-const traceRequiredServerFiles = async (publish: string): Promise<string[]> => {
-  const {
-    files,
-    relativeAppDir,
-    config: {
-      experimental: { outputFileTracingRoot },
-    },
-  } = await getRequiredServerFiles(publish)
-  const appDirRoot = join(outputFileTracingRoot, relativeAppDir)
-  const absoluteFiles = files.map((file) => join(appDirRoot, file))
-
-  absoluteFiles.push(join(publish, 'required-server-files.json'))
-
-  return absoluteFiles
-}
-
-const traceNextServer = async (publish: string): Promise<string[]> => {
-  const nextServerDeps = await getDependenciesOfFile(join(publish, 'next-server.js'))
-
-  // during testing, i've seen `next-server` contain only one line.
-  // this is a sanity check to make sure we're getting all the deps.
-  if (nextServerDeps.length < 10) {
-    console.error(nextServerDeps)
-    throw new Error("next-server.js.nft.json didn't contain all dependencies.")
-  }
-
-  const filtered = nextServerDeps.filter((f) => {
-    // NFT detects a bunch of large development files that we don't need.
-    if (f.endsWith('.development.js')) return false
-
-    // not needed for API Routes!
-    if (f.endsWith('node_modules/sass/sass.dart.js')) return false
-
-    return true
-  })
-
-  return filtered
-}
-
-export const traceNPMPackage = async (packageName: string, publish: string) => {
-  try {
-    return await glob(join(dirname(require.resolve(packageName, { paths: [publish] })), '**', '*'), {
-      absolute: true,
-    })
-  } catch (error) {
-    if (process.env.NODE_ENV === 'test') {
-      return []
-    }
-    throw error
-  }
-}
-
-export const getCommonDependencies = async (publish: string) => {
-  const deps = await Promise.all([
-    traceRequiredServerFiles(publish),
-    traceNextServer(publish),
-
-    // used by our own bridge.js
-    traceNPMPackage('follow-redirects', publish),
-
-    // used by bridge.js, only for ODB routes - they don't weigh a lot, so we can include them anyway
-    traceNPMPackage('@netlify/functions/package.json', publish),
-    traceNPMPackage('is-promise', publish),
-  ])
-
-  return deps.flat(1)
 }
 
 const sum = (arr: number[]) => arr.reduce((v, current) => v + current, 0)
